Try to fix labelInputField import error
