--- frontend/src/pages/Guests/NewGuest/LabelInputField.tsx
@@ -1,24 +1,0 @@
-import React from 'react';
-// Styled Components
-import { StyledTextField } from './styles';
-// Types
-import { FieldProps } from 'formik';
-
-export const labelInputField = (label: string) => {
-  return ({ field, form }: FieldProps) => {
-    const { name, value } = field;
-    const { touched, errors } = form;
-    const errorState = Boolean(errors[name] && touched[name]);
-    return (
-      <StyledTextField
-        error={errorState}
-        inputProps={{ ...field, 'data-testid': `input-${name}` }}
-        InputLabelProps={{ 'data-testid': `label-${name}` }}
-        className={`field-${name}`}
-        data-testid={`field-${name}`}
-        label={errorState ? errors[name] : label}
-        {...field}
-      />
-    );
-  };
-};
